action: use bash ansi c quoting for JSON args to curl
This _should_ help avoid issues where various types of quotes in `inputs.text` could cause problems when inserted into the JSON string we're building.

Mashing strings together and pretending it's JSON is never going to be perfect, but this should be _more_ resilient than plain single quotes or double quotes.

https://linux.die.net/man/1/bash#:~:text=Words%20of%20the%20form%20%24%27string%27%20are%20treated%20specially.